API adapter package reorganization
