+++ backend/src/main/java/club/klabis/adapters/api/ApiExceptionHandlers.java
@@ -12,31 +12,11 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
 @ControllerAdvice
-public class ApiExceptionHandlers {
-
-    @ExceptionHandler(MemberRegistrationFailedException.class)
-    public ErrorResponse handleMeberRegistrationError(MemberRegistrationFailedException error) {
-        return ErrorResponse.builder(error, ProblemDetail.forStatusAndDetail(HttpStatus.CONFLICT, error.getMessage())).build();
-    }
+class ApiExceptionHandlers {
 
     @ExceptionHandler(IncorrectFormDataException.class)
     public ErrorResponse handleMeberRegistrationError(IncorrectFormDataException error) {
         return ErrorResponse.builder(error, ProblemDetail.forStatusAndDetail(HttpStatus.BAD_REQUEST, error.getMessage())).build();
-    }
-
-    @ExceptionHandler(MemberNotFoundException.class)
-    public ErrorResponse handleMeberRegistrationError(MemberNotFoundException error) {
-        return ErrorResponse.builder(error, ProblemDetail.forStatusAndDetail(HttpStatus.NOT_FOUND, error.getMessage())).build();
-    }
-
-    @ExceptionHandler(MembershipCannotBeSuspendedException.class)
-    public ErrorResponse handleMembershipSuspensionFailed(MembershipCannotBeSuspendedException error) {
-        return ErrorResponse.builder(error, ProblemDetail.forStatusAndDetail(HttpStatus.CONFLICT, error.getMessage())).build();
-    }
-
-    @ExceptionHandler(ApplicationUserNotFound.class)
-    public ErrorResponse handleMembershipSuspensionFailed(ApplicationUserNotFound error) {
-        return ErrorResponse.builder(error, ProblemDetail.forStatusAndDetail(HttpStatus.NOT_FOUND, error.getMessage())).build();
     }
 
 }
