+++ backend/src/main/java/club/klabis/adapters/api/appusers/mappers/MemberGrantsFormMapper.java
@@ -1,4 +1,4 @@
-package club.klabis.adapters.api.mappers.appusers;
+package club.klabis.adapters.api.appusers.mappers;
 
 import club.klabis.api.dto.MemberGrantsFormApiDto;
 import club.klabis.common.mapstruct.DomainToDtoMapperConfiguration;
