+++ backend/src/main/java/club/klabis/adapters/api/members/MembersController.java
@@ -1,10 +1,9 @@
-package club.klabis.adapters.api;
+package club.klabis.adapters.api.members;
 
 import club.klabis.api.MembersApi;
 import club.klabis.api.dto.*;
 import club.klabis.domain.appusers.ApplicationGrant;
 import club.klabis.domain.appusers.ApplicationUser;
-import club.klabis.domain.appusers.ApplicationUserNotFound;
 import club.klabis.domain.appusers.ApplicationUserService;
 import club.klabis.domain.members.Member;
 import club.klabis.domain.members.MemberNotFoundException;
@@ -14,8 +13,6 @@ import org.springframework.core.convert.ConversionService;
 import org.springframework.core.convert.TypeDescriptor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.ErrorResponse;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Collection;
