+++ backend/src/main/java/club/klabis/adapters/api/members/mappers/AddressApiDtoMapper.java
@@ -1,4 +1,4 @@
-package club.klabis.adapters.api.mappers.members;
+package club.klabis.adapters.api.members.mappers;
 
 import club.klabis.api.dto.AddressApiDto;
 import club.klabis.common.mapstruct.DomainToDtoMapperConfiguration;
