+++ backend/src/main/java/club/klabis/adapters/api/members/mappers/LegalGuardianApiDtoMapper.java
@@ -1,9 +1,8 @@
-package club.klabis.adapters.api.mappers.members;
+package club.klabis.adapters.api.members.mappers;
 
-
-import club.klabis.api.dto.RefereeLicenceApiDto;
+import club.klabis.api.dto.LegalGuardianApiDto;
 import club.klabis.common.mapstruct.DomainToDtoMapperConfiguration;
-import club.klabis.domain.members.RefereeLicence;
+import club.klabis.domain.members.LegalGuardian;
 import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
@@ -11,13 +10,14 @@ import org.mapstruct.extensions.spring.DelegatingConverter;
 import org.springframework.core.convert.converter.Converter;
 
 @Mapper(config = DomainToDtoMapperConfiguration.class)
-interface RefereeLicenceApiDtoMapper extends Converter<RefereeLicence, RefereeLicenceApiDto> {
+interface LegalGuardianApiDtoMapper extends Converter<LegalGuardian, LegalGuardianApiDto> {
 
     @Override
-    @Mapping(target = "licence", source = "licenceType")
-    RefereeLicenceApiDto convert(RefereeLicence source);
+    @Mapping(source = "contacts", target = "contact")
+    LegalGuardianApiDto convert(LegalGuardian source);
 
     @DelegatingConverter
     @InheritInverseConfiguration
-    RefereeLicence fromApiDto(RefereeLicenceApiDto apiDto);
+    LegalGuardian fromApiDto(LegalGuardianApiDto apiDto);
+
 }
