+++ backend/src/main/java/club/klabis/adapters/api/members/mappers/MemberEditFormFromApiDtoMapper.java
@@ -1,17 +1,16 @@
-package club.klabis.adapters.api.mappers.members;
+package club.klabis.adapters.api.members.mappers;
 
-import club.klabis.api.dto.MemberRegistrationFormApiDto;
+import club.klabis.api.dto.MemberEditFormApiDto;
 import club.klabis.common.mapstruct.DomainToDtoMapperConfiguration;
-import club.klabis.domain.members.forms.RegistrationForm;
+import club.klabis.domain.members.forms.MemberEditForm;
 import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
 import org.mapstruct.extensions.spring.DelegatingConverter;
 import org.springframework.core.convert.converter.Converter;
 
 @Mapper(config = DomainToDtoMapperConfiguration.class)
-interface MemberRegistrationFormApiDtoMapper extends Converter<RegistrationForm, MemberRegistrationFormApiDto> {
+interface MemberEditFormFromApiDtoMapper extends Converter<MemberEditForm, MemberEditFormApiDto> {
     @DelegatingConverter
     @InheritInverseConfiguration
-    RegistrationForm reverseConvert(MemberRegistrationFormApiDto apiDto);
-
+    MemberEditForm reverseConvert(MemberEditFormApiDto apiDto);
 }
