+++ backend/src/main/java/club/klabis/adapters/api/members/mappers/MemberRegistrationFormApiDtoMapper.java
@@ -1,16 +1,17 @@
-package club.klabis.adapters.api.mappers.members;
+package club.klabis.adapters.api.members.mappers;
 
-import club.klabis.api.dto.SexApiDto;
+import club.klabis.api.dto.MemberRegistrationFormApiDto;
 import club.klabis.common.mapstruct.DomainToDtoMapperConfiguration;
-import club.klabis.domain.members.Sex;
+import club.klabis.domain.members.forms.RegistrationForm;
 import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
 import org.mapstruct.extensions.spring.DelegatingConverter;
 import org.springframework.core.convert.converter.Converter;
 
 @Mapper(config = DomainToDtoMapperConfiguration.class)
-interface SexApiDtoMapper extends Converter<Sex, SexApiDto> {
+interface MemberRegistrationFormApiDtoMapper extends Converter<RegistrationForm, MemberRegistrationFormApiDto> {
     @DelegatingConverter
     @InheritInverseConfiguration
-    Sex fromApiDto(SexApiDto apiDto);
+    RegistrationForm reverseConvert(MemberRegistrationFormApiDto apiDto);
+
 }
