+++ backend/src/main/java/club/klabis/adapters/api/members/mappers/RefereeLicenceApiDtoMapper.java
@@ -1,8 +1,9 @@
-package club.klabis.adapters.api.mappers.members;
+package club.klabis.adapters.api.members.mappers;
 
-import club.klabis.api.dto.LegalGuardianApiDto;
+
+import club.klabis.api.dto.RefereeLicenceApiDto;
 import club.klabis.common.mapstruct.DomainToDtoMapperConfiguration;
-import club.klabis.domain.members.LegalGuardian;
+import club.klabis.domain.members.RefereeLicence;
 import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
@@ -10,14 +11,13 @@ import org.mapstruct.extensions.spring.DelegatingConverter;
 import org.springframework.core.convert.converter.Converter;
 
 @Mapper(config = DomainToDtoMapperConfiguration.class)
-interface LegalGuardianApiDtoMapper extends Converter<LegalGuardian, LegalGuardianApiDto> {
+interface RefereeLicenceApiDtoMapper extends Converter<RefereeLicence, RefereeLicenceApiDto> {
 
     @Override
-    @Mapping(source = "contacts", target = "contact")
-    LegalGuardianApiDto convert(LegalGuardian source);
+    @Mapping(target = "licence", source = "licenceType")
+    RefereeLicenceApiDto convert(RefereeLicence source);
 
     @DelegatingConverter
     @InheritInverseConfiguration
-    LegalGuardian fromApiDto(LegalGuardianApiDto apiDto);
-
+    RefereeLicence fromApiDto(RefereeLicenceApiDto apiDto);
 }
