+++ backend/src/main/java/club/klabis/adapters/api/members/mappers/SexApiDtoMapper.java
@@ -1,16 +1,16 @@
-package club.klabis.adapters.api.mappers.members;
+package club.klabis.adapters.api.members.mappers;
 
-import club.klabis.api.dto.MemberEditFormApiDto;
+import club.klabis.api.dto.SexApiDto;
 import club.klabis.common.mapstruct.DomainToDtoMapperConfiguration;
-import club.klabis.domain.members.forms.MemberEditForm;
+import club.klabis.domain.members.Sex;
 import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
 import org.mapstruct.extensions.spring.DelegatingConverter;
 import org.springframework.core.convert.converter.Converter;
 
 @Mapper(config = DomainToDtoMapperConfiguration.class)
-interface MemberEditFormFromApiDtoMapper extends Converter<MemberEditForm, MemberEditFormApiDto> {
+interface SexApiDtoMapper extends Converter<Sex, SexApiDto> {
     @DelegatingConverter
     @InheritInverseConfiguration
-    MemberEditForm reverseConvert(MemberEditFormApiDto apiDto);
+    Sex fromApiDto(SexApiDto apiDto);
 }
